Task 13

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -21,8 +21,9 @@ if __name__ == "__main__":
 
     states = session.query(State).filter(State.name.like("%a%")).all()
 
-    for row in states:
-        session.delete(row)
+    if states:
+        for row in states:
+            session.delete(row)
 
     session.commit()
